fix paramater type

+++ print_pointer.c
@@ -6,7 +6,7 @@
 /*   By: jestevam < jestevam@student.42sp.org.br    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/29 14:45:05 by jestevam          #+#    #+#             */
-/*   Updated: 2021/07/06 19:27:45 by jestevam         ###   ########.fr       */
+/*   Updated: 2021/07/06 20:31:19 by jestevam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ static void	print_pointer(unsigned long int numb, int s_dot)
 			ft_putchar_fd('0', 1);
 	}
 	else
-		count_print_pointer(numb, BASE_LOWER_HEXA, 1);
+		count_print_pointer(numb, BASE_LOWER_HEXA, 1);	
 }
 
 static void	print_r_l(int dif, unsigned long int numb, int sinal, int s_dot)
@@ -51,7 +51,7 @@ static void	print_r_l(int dif, unsigned long int numb, int sinal, int s_dot)
 	}
 }
 
-static void	align_print(int len_num, int numb, t_flags *flag)
+static void	align_print(int len_num, unsigned long int numb, t_flags *flag)
 {
 	if (flag->width > len_num)
 	{
